issue #525 onnx refactor
