+++ openeo-geotrellis/src/test/scala/org/openeo/geotrellis/OpenEOProcessesSpec.scala
@@ -1030,7 +1030,7 @@ class OpenEOProcessesSpec extends RasterMatchers {
         if (path.startsWith("http")) path
         else getClass.getResource(path).getPath
       val datacube = TileLayerRDDBuilders.createMultibandTileLayerRDD(OpenEOProcessesSpec.sc, tile, new TileLayout(layoutCols, layoutRows, tileSize, tileSize))
-      val resultCube = new OpenEOProcesses().predictONNXSpatial(datacube,model)
+      val resultCube = new OpenEOProcesses().predictONNXGeneric(datacube,model)
       assertEquals(expectedType, resultCube.metadata.cellType)
       val theResultTile = resultCube.stitch().tile
       assertEquals(expectedNBands,theResultTile.bandCount)
@@ -1123,7 +1123,7 @@ class OpenEOProcessesSpec extends RasterMatchers {
       val model =
         if (path.startsWith("http")) path
        else getClass.getResource(path).getPath
-      val resultCube = new OpenEOProcesses().predictONNXTemporal(datacube,model)
+      val resultCube = new OpenEOProcesses().predictONNXGeneric(datacube,model)
       assertEquals(expectedType, resultCube.metadata.cellType)
 
       val results = resultCube.toSpatial(date)
